chore: fix IE11 bundle
