+++ packages/fiori/bundle.common.js
@@ -1,0 +1,21 @@
+// FIORI features
+import "./dist/features/CoPilotAnimation.js";
+
+// FIORI components
+import Bar from "./dist/Bar.js";
+import FlexibleColumnLayout from "./dist/FlexibleColumnLayout.js";
+import ProductSwitch from "./dist/ProductSwitch.js";
+import ProductSwitchItem from "./dist/ProductSwitchItem.js";
+import SideNavigation from "./dist/SideNavigation.js";
+import SideNavigationItem from "./dist/SideNavigationItem.js";
+import SideNavigationSubItem from "./dist/SideNavigationSubItem.js";
+import Page from "./dist/Page.js";
+import ShellBar from "./dist/ShellBar.js";
+import ShellBarItem from "./dist/ShellBarItem.js";
+import Timeline from "./dist/Timeline.js";
+import UploadCollection from "./dist/UploadCollection.js";
+import UploadCollectionItem from "./dist/UploadCollectionItem.js";
+import NotificationListItem from "./dist/NotificationListItem.js"
+import NotificationListGroupItem from "./dist/NotificationListGroupItem.js";
+import NotificationAction from "./dist/NotificationAction.js";
+import Wizard from "./dist/Wizard.js";
